requirements.txt
torch
torchvision
numpy
opencv-python
pillow
scikit-image
transformers
sentencepiece
matplotlib
tqdm
python-dotenv
diffusers
accelerate
hf_xet
scikit-learn
pandas
seaborn
tabulate
kagglehub